New (useful) life for this repo!
Basically, the CLUI is dead. It was a futile attempt to make a
GUI on the CL, made in the hardest possible way. So, I deleted it.

After some time of intensive use of the terminal, I came to realize
that my biggest pain point was the fear of changing the configs of
the programs I use and losing all my hacky scripts. I work on a
Chromebook, and my dear configs and scripts are only 1 powerwash
away of disappearing (and it has already happened once).

So, this script was born. In this way, I can do clean Linux installs/
Chrome OS powerwashes fearlessly.

For now, my customizations are:
	- bash
	- git
	- nano
	- tmux

On branch master

Changes to be committed:
	deleted:    CLUI/CLUI.sh
	deleted:    CLUI/k
	modified:   README.md
	deleted:    Script.sh

	new file:   config_installer.sh
	new file:   tmux_stuff.zip
	new file:   dot_files/bashrc
	new file:   dot_files/gitconfig
	new file:   dot_files/nanorc
	new file:   dot_files/tmux.conf