Removed useless file: ./separator
Now, the variable $CUSTOM_BAR, that was moved to the file config.conf
on commit ba24fae, contains the line instead of a file that contains
the line. Also, in README there where a lot of references to a
$SEPARATOR - that is the $CUSTOM_BAR. The README was updated
accordingly, and so where config_installer.sh and
config_uninstaller.sh.

On branch master

Changes to be committed:
	deleted:    separator
	modified:   config.conf
	modified:   README.md
	modified:   config_installer.sh
	modified:   config_uninstaller.sh