Now using a truly random temporary file
As commit a653ab3 says, a temp file is needed during the process of
reverting configs to a previous state/removing the appended lines.

The temp file was not a "real" temp file - it was a deterministic
file with a weird name (r4nd0m_f1l3n4m3). Now, we use mktemp, part
of the coreutils, that creates a real temp file with a random file
name.

On branch master

Changes to be committed:
	modified:   README.md
	modified:   config.conf