Add support for more types of config files
+ Add `set -e` to all the scripts
+ Remove check to delete files if scripts are equal
+ Add verification that `$SCRIPTS_PATH` exists
+ Add vim config
+ Auto-tmux: decrease time in watch command for mini-map
+ Update README

---

Previously, I was relying in `#` introducing a comment line in configs.
Unfortunately, that is not true for all configs -- vim uses `"`. So, to
allow this and other situations, instead of specifying $CUSTOM_BAR,
the separator is calculated on a file-by-file basis, using the shell
function `separator_calc` -- basically, grabs the first line and repeats
32 times (currently).

So, now all the configs must have a first line empty, except for the
comment character.

Changes to be committed:
	modified:   README.md
	modified:   config.conf
	modified:   config_installer.sh
	modified:   config_uninstaller.sh
	modified:   dot_files/bashrc
	modified:   dot_files/gitconfig
	modified:   dot_files/nanorc
	modified:   dot_files/tmux.conf
	new file:   dot_files/vimrc
	modified:   tmux_stuff/Auto-tmux.sh